[po-196] Code review, moved the competence result computation from CampaignParticipation to CampaignParticipationResult
